painting-service/requirements.txt

Flask-Migrate==3.1.0
Flask-WTF==0.15.1
WTForms==2.3.3
gunicorn==20.1.0
pymysql==1.0.2
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Jinja2==3.1.2